Rename treeView.ts -> checkoutFlexTreeView.ts.
